Move Victory-specific classes into seperate package
