+++ org.emoflon.ibex.tgg.core.runtime/src/org/emoflon/ibex/tgg/operational/victory/IVictoryDataProvider.java
@@ -1,4 +1,4 @@
-package org.emoflon.ibex.tgg.operational.monitoring;
+package org.emoflon.ibex.tgg.operational.victory;
 
 import java.io.IOException;
 import java.util.Collection;
